JI:693902 (MEAPCLIENT-738)
Declaring an undefined property in a type declaration throws an error

When creating a private scope, the proper isPropertyDescriptor check
was not run, and property descriptors were not properly added.
This change correctly uses the isPropertyDescriptor function, and
makes it behave more correctly.

Note: must bump npm build version every commit for now

Change-Id: Ic5c2562c41b9be8c92983dd79f29de5cbcbf9904